+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.0.2
 * Fix posts/all?hashes arg validation
+* Add unknown argument validation
 
 ### 0.0.1
 * Initial version